+++ web/src/app/object.rs
@@ -1,10 +1,7 @@
 use core::marker::PhantomData;
 
-use xitca_http::util::service::router::IntoObject;
-use xitca_service::{
-    object::{BoxedSyncServiceObject, ServiceObject},
-    Service,
-};
+use xitca_http::util::service::router::{IntoObject, PathGen, RouteGen, RouteObject};
+use xitca_service::{object::ServiceObject, Service};
 
 use crate::context::WebContext;
 
@@ -14,17 +11,37 @@ impl<C, B, I, Res, Err> IntoObject<I, ()> for WebContext<'_, C, B>
 where
     C: 'static,
     B: 'static,
-    I: Service + Send + Sync + 'static,
+    I: Service + RouteGen + Send + Sync + 'static,
     I::Response: for<'r> Service<WebContext<'r, C, B>, Response = Res, Error = Err> + 'static,
 {
-    type Object = BoxedSyncServiceObject<(), WebObject<C, B, Res, Err>, I::Error>;
+    type Object = RouteObject<(), WebObject<C, B, Res, Err>, I::Error>;
 
     fn into_object(inner: I) -> Self::Object {
         struct Builder<I, C, B>(I, PhantomData<fn(C, B)>);
 
+        impl<I, C, B> PathGen for Builder<I, C, B>
+        where
+            I: PathGen,
+        {
+            fn path_gen(&mut self, prefix: String) -> String {
+                self.0.path_gen(prefix)
+            }
+        }
+
+        impl<I, C, B> RouteGen for Builder<I, C, B>
+        where
+            I: RouteGen,
+        {
+            type Route<R> = I::Route<R>;
+
+            fn route_gen<R>(route: R) -> Self::Route<R> {
+                I::route_gen(route)
+            }
+        }
+
         impl<C, I, B, Res, Err> Service for Builder<I, C, B>
         where
-            I: Service + 'static,
+            I: Service + RouteGen + 'static,
             I::Response: for<'r> Service<WebContext<'r, C, B>, Response = Res, Error = Err> + 'static,
         {
             type Response = WebObject<C, B, Res, Err>;
@@ -35,6 +52,6 @@ where
             }
         }
 
-        Box::new(Builder(inner, PhantomData))
+        RouteObject(Box::new(Builder(inner, PhantomData)))
     }
 }